+++ app/src/main/res/layout/activity_cart.xml
@@ -31,7 +31,7 @@
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragment_container_view_cart"
             android:layout_width="0dp"
-            android:layout_height="wrap_content"
+            android:layout_height="0dp"
             app:layout_constraintBottom_toTopOf="@id/layout_order_bottom_bar"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
